Merged two passes into a single pass.
Rather than first making sure all of the
stabilizers commute with op and then filtering
out the identities, I realized it would be more
efficient just to make a single pass and in this
pass filter out all stabilizers that are equal
to op's conjugal partner.  This also has the
advantage of being closer to the theorem that
is being written up in the paper.