Rewrote logical qubit computing algorithm.
I made significant modifications of the algorithm
so that eliminating the redundant stabilizers and
computing the logical qubits now share the same
gaussian elimination routine.  I then rewrote
the code which computes the logical qubits so that
it uses a simpler strategy, using information
that is now available from the gaussian
elimination rather than forcing the generators
into "standard form".