Eliminated codelib.cc;  now the library consists of the header file alone.
In the process of doing this, I also fixed the flaw in the library where the <<
operator was defined for dynamic qubits and arbitrary codes, but *not* for
static qubits, causing the user to have to manually implement this operator
so that the qec << operator could reference it.  Now << is implemented for all
quantum operators, and as a bonus operators now have a "to_string()" method that
creates a string representation of the operator.